refactor: reorganize type exports and enhance type definitions in core package
