+++ packages/core/src/index.ts
@@ -3,3 +3,12 @@ export {
   withPaletteNumber,
   getSafeThemeColor,
 } from './utils/colors-utils'
+
+export type {
+  ThemeColor,
+  EdgeInsets,
+  Size,
+  Radius,
+  SizeValues,
+  RadiusValues,
+} from './types'
