+++ packages/core/src/types/edge-insets.ts
@@ -1,0 +1,9 @@
+export type EdgeInsets = {
+  top?: number
+  bottom?: number
+  left?: number
+  right?: number
+  horizontal?: number
+  vertical?: number
+  all?: number
+}
