+++ packages/core/src/types/index.ts
@@ -1,0 +1,3 @@
+export type { ThemeColor } from './theme'
+export type { EdgeInsets } from './edge-insets'
+export type { Size, Radius, SizeValues, RadiusValues } from './sizing'
