+++ packages/core/src/types/sizing.ts
@@ -1,0 +1,18 @@
+export type Size = 'xs' | 'sm' | 'md' | 'lg'
+
+export type Radius = 'none' | 'sm' | 'md' | 'lg' | 'full'
+
+export type SizeValues = {
+  xs: number
+  sm: number
+  md: number
+  lg: number
+}
+
+export type RadiusValues = {
+  none: number
+  sm: number
+  md: number
+  lg: number
+  full: number
+}
