+++ packages/hybrid/src/types/index.ts
@@ -1,20 +1,8 @@
-export * from './theme'
-
-export type Size = 'xs' | 'sm' | 'md' | 'lg'
-
-export type Radius = 'none' | 'sm' | 'md' | 'lg' | 'full'
-
-export type SizeValues = {
-  xs: number
-  sm: number
-  md: number
-  lg: number
-}
-
-export type RadiusValues = {
-  none: number
-  sm: number
-  md: number
-  lg: number
-  full: number
-}
+export type {
+  ThemeColor,
+  Size,
+  Radius,
+  SizeValues,
+  RadiusValues,
+  EdgeInsets,
+} from '@xaui/core/types'
