--- packages/native/src/types/theme.ts
@@ -1,8 +1,0 @@
-export type ThemeColor =
-  | 'primary'
-  | 'secondary'
-  | 'tertiary'
-  | 'danger'
-  | 'warning'
-  | 'success'
-  | 'default'
